fix: missing filter index over the primary keys
